[new-suggestions] - Refactors UI and layout, and also refactor the components to increase readability
