+++ src/components/generic-link-share/component.tsx
@@ -256,7 +256,6 @@ function GenericLinkShare(
       pluginApi.setGenericContentItems([]);
     }
   }, [link]);
-
   return (
     <ModalToShareLink
       {...{
