+++ src/components/modal-to-share-link/link-form/checkbox-input/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { FormToSendUrlItem as FormToSendUrlItemCommon } from '../styles';
+
+export const FormToSendUrlItem = styled(FormToSendUrlItemCommon)``;
+
+export const LabelFormCheckbox = styled.span`
+  margin-left: 10px;
+`;
