+++ src/components/modal-to-share-link/link-form/checkbox-input/types.ts
@@ -1,0 +1,4 @@
+export interface CheckboxInputComponentProps {
+  isUrlSameForRole: boolean;
+  handleCheckboxChange: () => void;
+}
