+++ src/components/modal-to-share-link/link-form/styles.ts
@@ -2,14 +2,14 @@ import styled, { css } from 'styled-components';
 import { ButtonStyleProps, FormToSendUrlItemProps } from './types';
 
 const colorGray = 'var(--color-gray, #4E5A66)';
+const borderSizeSmall = '1px';
 const borderSize = '2px';
+const colorGrayLightest = 'var(--color-gray-lightest, #D4D9DF)';
+const lineHeightComputed = '1rem';
+const mdPaddingX = '1rem';
 
-export const LabelFormTextInput = styled.input`
-  width: 100%;
-
-  &::placeholder {
-    color: #555;
-  }
+export const ButtonsWrapper = styled.div`
+  align-self: flex-end;
 `;
 
 export const ButtonStyle = styled.button<ButtonStyleProps>`
@@ -71,8 +71,60 @@ export const ButtonStyle = styled.button<ButtonStyleProps>`
   }
 `;
 
-export const LabelForm = styled.label`
-  margin-right: 5px;
+export const WarningIframeMessage = styled.p`
+  color: #4e5a66;
+  font-size: 0.875rem;
+  font-style: italic;
+  margin: .5rem 0 0 0;
+`;
+
+export const SendingButton = styled.input`
+  border: 3px solid transparent;
+  overflow: visible;
+  display: inline-block;
+  border-radius: 2px;
+  font-weight: 600;
+  line-height: 1;
+  text-align: center;
+  white-space: nowrap;
+  vertical-align: middle;
+  cursor: pointer;
+  user-select: none;
+
+  &:hover {
+    border: 3px solid #0f70d7;
+  }
+  padding: 0.45rem 1rem;
+  background-color: #0f70d7;
+  color: #fff;
+  font-size: calc(1rem * 0.85);
+`;
+
+export const FormToSendUrl = styled.form`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  height: 100%;
+  width: 85%;
+`;
+
+export const FormWrapper = styled.div`
+  display: flex;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-items: center;
+`;
+
+export const FormBody = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-Left: .5rem;
+  width: 100%;
+  height: 100%;
+  align-items: flex-start;
+  justify-items: flex-start;
 `;
 
 export const FormToSendUrlItem = styled.div<FormToSendUrlItemProps>`
@@ -88,4 +140,13 @@ export const FormToSendUrlItem = styled.div<FormToSendUrlItemProps>`
       `
   )}
   width: 100%;
+`;
+
+export const BottomSeparator = styled.div`
+  position: relative;
+  width: 100%;
+  height: ${borderSizeSmall};
+  align-self: center;
+  background-color: ${colorGrayLightest};
+  margin: calc(${lineHeightComputed} * 2) ${mdPaddingX} calc(${lineHeightComputed} * 1.75) ${mdPaddingX};
 `;
